Slice orb logic into its own file
